create profile service

+++ domain/dtos/response/user_profile_response.go
@@ -4,8 +4,8 @@ import "github.com/ariwiraa/my-gram/domain"
 
 type UserProfileResponse struct {
 	Username   string         `json:"username"`
-	PostsCount uint           `json:"posts_count"`
-	Follower   uint           `json:"follower"`
-	Following  uint           `json:"following"`
+	PostsCount int64          `json:"posts_count"`
+	Follower   int64          `json:"follower"`
+	Following  int64          `json:"following"`
 	Posts      []domain.Photo `json:"posts"`
 }
